Dev: Code refactored: mock tests renamed and grouped; readme upgraded.
