+++ README.md
@@ -38,6 +38,7 @@ electronic notes by the Spring framework.
 + 3.1.1 UserServiceTest class, UserTestData implementation (completed), initDB.sql, populateDB.sql (completed);
 + 3.1.2 NotepadServiceTest class, NotepadTestData implementation (completed), initDB.sql, populateDB.sql (competed);
 + 3.1.3 NoteServiceTest class, NoteTestData implementation (completed), initDB.sql, populateDB.sql (competed);
++ 3.1.4 TagServiceTest class, TagTestData implementation (completed), initDB.sql, populateDB.sql (competed);
 + 3.2 ServiceValidatorUtilTest (completed);
 + 3.3 Mock test Services(completed);
 ```
